requirements.txt
Flask==3.0.3
google-generativeai>=0.7.2
xai-sdk==1.5.0
requests==2.32.5
